requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==3.0.2
msgspec==0.18.6
python-magic-bin==0.4.14
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
